requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cryptography==45.0.4
cssselect2==0.8.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
google-auth==2.40.3
html5lib==1.1
idna==3.10
lxml==5.4.0
openpyxl==3.1.5
oscrypto==1.3.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
pypdf==5.6.0
python-bidi==0.6.6
python-decouple==3.8
PyYAML==6.0.2
razorpay==1.4.2
reportlab==4.4.1
requests==2.32.3
rsa==4.9.1
setuptools==80.7.1
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
webencodings==0.5.1
xhtml2pdf==0.2.17
